Added limited handling of NOT in CanonizerService (Closes #5).
The CanonizerService will now push `NOT` operators into expressions of the form `A OP B`, where `OP` is one of `EQ`, `NE`, `GE`, `GT`, `LE`, `LT`.  To be more complete, the canonizer should really push NOT into all kinds of expressions, but the current fix will at least the specific problem mentioned in #5.